requirements.txt
Django==5.2.5
python-decouple
psycopg[binary]
gunicorn